+++ pay-api/src/pay_api/models/payment.py
@@ -27,10 +27,11 @@ from sqlalchemy.sql import column, select, text, table
 
 from pay_api.exceptions import BusinessException
 from pay_api.utils.constants import DT_SHORT_FORMAT
-from pay_api.utils.enums import DatabaseViews, TransactionsViewColumns, InvoiceReferenceStatus
+from pay_api.utils.enums import DatabaseViews, InvoiceReferenceStatus
 from pay_api.utils.enums import PaymentMethod as PaymentMethodEnum
 from pay_api.utils.enums import PaymentStatus
 from pay_api.utils.errors import Error
+from pay_api.utils.query_helpers import TransactionQuery
 from pay_api.utils.serializable import Serializable
 from pay_api.utils.user_context import UserContext, user_context
 from pay_api.utils.util import get_first_and_last_dates_of_month, get_str_by_path, get_week_start_and_end_date
@@ -252,7 +253,7 @@ class Payment(BaseModel):  # pylint: disable=too-many-instance-attributes
     @classmethod
     def get_today_invoices(cls):
         """Fetch invoices created on today's date, based on the base query structure."""
-        base_query = Serializable.generate_base_transaction_query()
+        base_query = TransactionQuery.generate_base_transaction_query()
 
         today_query = (
             base_query
@@ -283,7 +284,7 @@ class Payment(BaseModel):  # pylint: disable=too-many-instance-attributes
             MetaData(),
             autoload_with=db.engine
         )
-        view_columns = [col.name for col in transactions_materialized_view.columns]
+        # view_columns = [col.name for col in transactions_materialized_view.columns]
         # base_query = Serializable.generate_base_transaction_query()
         # today_query = base_query.with_entities(
         #     *(getattr(Invoice, col) if hasattr(Invoice, col) else literal_column("NULL").label(col) for col in view_columns)
@@ -303,16 +304,74 @@ class Payment(BaseModel):  # pylint: disable=too-many-instance-attributes
 
         # Create a query using the subquery
         # query = db.session.query(historical_subquery)
+        
         query = db.session.query(transactions_materialized_view)
         query = cls.filter(query, auth_account_id, search_filter, table_alias=transactions_materialized_view)
+        # query = (
+        #     db.session.query(Invoice)
+        #     .outerjoin(PaymentAccount, Invoice.payment_account_id == PaymentAccount.id)
+        #     .outerjoin(PaymentLineItem, PaymentLineItem.invoice_id == Invoice.id)
+        #     .outerjoin(
+        #         FeeSchedule,
+        #         FeeSchedule.fee_schedule_id == PaymentLineItem.fee_schedule_id,
+        #     )
+        #     .outerjoin(InvoiceReference, InvoiceReference.invoice_id == Invoice.id)
+        #     .options(
+        #         lazyload("*"),
+        #         load_only(
+        #             Invoice.id,
+        #             Invoice.corp_type_code,
+        #             Invoice.created_on,
+        #             Invoice.payment_date,
+        #             Invoice.refund_date,
+        #             Invoice.invoice_status_code,
+        #             Invoice.total,
+        #             Invoice.service_fees,
+        #             Invoice.paid,
+        #             Invoice.refund,
+        #             Invoice.folio_number,
+        #             Invoice.created_name,
+        #             Invoice.invoice_status_code,
+        #             Invoice.payment_method_code,
+        #             Invoice.details,
+        #             Invoice.business_identifier,
+        #             Invoice.created_by,
+        #             Invoice.filing_id,
+        #             Invoice.bcol_account,
+        #             Invoice.disbursement_date,
+        #             Invoice.disbursement_reversal_date,
+        #             Invoice.overdue_date,
+        #         ),
+        #         contains_eager(Invoice.payment_line_items)
+        #         .load_only(
+        #             PaymentLineItem.description,
+        #             PaymentLineItem.gst,
+        #             PaymentLineItem.pst,
+        #         )
+        #         .contains_eager(PaymentLineItem.fee_schedule)
+        #         .load_only(FeeSchedule.filing_type_code),
+        #         contains_eager(Invoice.payment_account).load_only(
+        #             PaymentAccount.auth_account_id,
+        #             PaymentAccount.name,
+        #             PaymentAccount.billable,
+        #         ),
+        #         contains_eager(Invoice.references).load_only(
+        #             InvoiceReference.invoice_number,
+        #             InvoiceReference.reference_number,
+        #             InvoiceReference.status_code,
+        #         ),
+        #     )
+        # )
+        # query = cls.filter(query, auth_account_id, search_filter)
         print("Query2:", str(query))
         if not return_all:
             count = cls.get_count(auth_account_id, search_filter)
             # Add pagination
             sub_query = cls.generate_subquery(auth_account_id, search_filter, limit, page)
+            # result_query = query.order_by(Invoice.id.desc()).filter(Invoice.id.in_(sub_query.subquery().select()))
             result_query = (
-                query.filter(transactions_materialized_view.c.id.in_(sub_query.subquery().select()))
-                .order_by(transactions_materialized_view.c.id.desc())
+                query.filter(transactions_materialized_view.c.invoice_id.in_(sub_query.subquery().select()))
+                .order_by(transactions_materialized_view.c.invoice_id.desc())
                 .execution_options(stream_results=True)
             )
             print("result_query<<<<<<<<<<<<<<:", str(result_query.statement.compile(compile_kwargs={"literal_binds": True})))
